+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented here.
 
-##  0.4.1 [Unreleased]
+##  0.4.1 [2026-05-31]
 
 ### Added